tests: Playing with Carbon Default is dangerous.
These variable are global. It has consequences.

People launching tests another timezone where facing random issues.
It was random because the tests are launched in different order on each run.

This was hidden by the CI forcing the timezone to PRC, and impossible to see by
someone working in PRC timezone.